requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.1
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
